Support for better final hit splits + Zang end fix
Fixed autosplit for the ending of Zangetsu mode to be on the final hit rather than 5 seconds later, using a new method of checking for coordinate boxes, player HP and boss HP. This boss HP pointer works for most if not all bosses, so this same setup may be used for other bosses. This change also affects Gebel's fight, since it uses the same arena. Excluded this coordinate box from the previous autosplitting method as well, to prevent double splits.